+++ mz_strm.h
@@ -34,10 +34,10 @@ extern "C" {
 
 typedef int32_t (*mz_stream_open_cb)(void *stream, const char *path, int32_t mode);
 typedef int32_t (*mz_stream_is_open_cb)(void *stream);
-typedef int32_t (*mz_stream_read_cb)(void *stream, void *buf, int32_t size);
-typedef int32_t (*mz_stream_write_cb)(void *stream, const void *buf, int32_t size);
+typedef int64_t (*mz_stream_read_cb)(void *stream, void *buf, int64_t size);
+typedef int64_t (*mz_stream_write_cb)(void *stream, const void *buf, int64_t size);
 typedef int64_t (*mz_stream_tell_cb)(void *stream);
-typedef int32_t (*mz_stream_seek_cb)(void *stream, int64_t offset, int32_t origin);
+typedef int64_t (*mz_stream_seek_cb)(void *stream, int64_t offset, int32_t origin);
 typedef int32_t (*mz_stream_close_cb)(void *stream);
 typedef int32_t (*mz_stream_error_cb)(void *stream);
 typedef void *(*mz_stream_create_cb)(void);
@@ -76,13 +76,13 @@ typedef struct mz_stream_s {
 
 int32_t mz_stream_open(void *stream, const char *path, int32_t mode);
 int32_t mz_stream_is_open(void *stream);
-int32_t mz_stream_read(void *stream, void *buf, int32_t size);
+int64_t mz_stream_read(void *stream, void *buf, int64_t size);
 int32_t mz_stream_read_uint8(void *stream, uint8_t *value);
 int32_t mz_stream_read_uint16(void *stream, uint16_t *value);
 int32_t mz_stream_read_uint32(void *stream, uint32_t *value);
 int32_t mz_stream_read_int64(void *stream, int64_t *value);
 int32_t mz_stream_read_uint64(void *stream, uint64_t *value);
-int32_t mz_stream_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_write(void *stream, const void *buf, int64_t size);
 int32_t mz_stream_write_uint8(void *stream, uint8_t value);
 int32_t mz_stream_write_uint16(void *stream, uint16_t value);
 int32_t mz_stream_write_uint32(void *stream, uint32_t value);
@@ -113,10 +113,10 @@ void mz_stream_delete(void **stream);
 
 int32_t mz_stream_raw_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_raw_is_open(void *stream);
-int32_t mz_stream_raw_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_raw_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_raw_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_raw_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_raw_tell(void *stream);
-int32_t mz_stream_raw_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_raw_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_raw_close(void *stream);
 int32_t mz_stream_raw_error(void *stream);
