+++ mz_strm_buf.h
@@ -21,10 +21,10 @@ extern "C" {
 
 int32_t mz_stream_buffered_open(void *stream, const char *path, int32_t mode);
 int32_t mz_stream_buffered_is_open(void *stream);
-int32_t mz_stream_buffered_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_buffered_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_buffered_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_buffered_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_buffered_tell(void *stream);
-int32_t mz_stream_buffered_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_buffered_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_buffered_close(void *stream);
 int32_t mz_stream_buffered_error(void *stream);
