+++ mz_strm_lzma.c
@@ -144,7 +144,7 @@ int32_t mz_stream_lzma_is_open(void *stream) {
     return MZ_OK;
 }
 
-int32_t mz_stream_lzma_read(void *stream, void *buf, int32_t size) {
+int64_t mz_stream_lzma_read(void *stream, void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_DECOMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -308,7 +308,7 @@ static int32_t mz_stream_lzma_code(void *stream, int32_t flush) {
 }
 #endif
 
-int32_t mz_stream_lzma_write(void *stream, const void *buf, int32_t size) {
+int64_t mz_stream_lzma_write(void *stream, const void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_COMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -337,7 +337,7 @@ int64_t mz_stream_lzma_tell(void *stream) {
     return MZ_TELL_ERROR;
 }
 
-int32_t mz_stream_lzma_seek(void *stream, int64_t offset, int32_t origin) {
+int64_t mz_stream_lzma_seek(void *stream, int64_t offset, int32_t origin) {
     MZ_UNUSED(stream);
     MZ_UNUSED(offset);
     MZ_UNUSED(origin);
