+++ mz_strm_lzma.h
@@ -19,10 +19,10 @@ extern "C" {
 
 int32_t mz_stream_lzma_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_lzma_is_open(void *stream);
-int32_t mz_stream_lzma_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_lzma_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_lzma_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_lzma_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_lzma_tell(void *stream);
-int32_t mz_stream_lzma_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_lzma_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_lzma_close(void *stream);
 int32_t mz_stream_lzma_error(void *stream);
