+++ mz_strm_mem.h
@@ -19,19 +19,19 @@ extern "C" {
 
 int32_t mz_stream_mem_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_mem_is_open(void *stream);
-int32_t mz_stream_mem_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_mem_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_mem_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_mem_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_mem_tell(void *stream);
-int32_t mz_stream_mem_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_mem_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_mem_close(void *stream);
 int32_t mz_stream_mem_error(void *stream);
 
-void mz_stream_mem_set_buffer(void *stream, void *buf, int32_t size);
+void mz_stream_mem_set_buffer(void *stream, void *buf, int64_t size);
 int32_t mz_stream_mem_get_buffer(void *stream, const void **buf);
 int32_t mz_stream_mem_get_buffer_at(void *stream, int64_t position, const void **buf);
 int32_t mz_stream_mem_get_buffer_at_current(void *stream, const void **buf);
-void mz_stream_mem_get_buffer_length(void *stream, int32_t *length);
-void mz_stream_mem_set_buffer_limit(void *stream, int32_t limit);
+void mz_stream_mem_get_buffer_length(void *stream, int64_t *length);
+void mz_stream_mem_set_buffer_limit(void *stream, int64_t limit);
 void mz_stream_mem_set_grow_size(void *stream, int32_t grow_size);
 
 void *mz_stream_mem_create(void);
