+++ mz_strm_ppmd.c
@@ -304,7 +304,7 @@ int32_t mz_stream_ppmd_is_open(void *stream) {
     return MZ_OK;
 }
 
-int32_t mz_stream_ppmd_read(void *stream, void *buf, int32_t size) {
+int64_t mz_stream_ppmd_read(void *stream, void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_DECOMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -356,7 +356,7 @@ int32_t mz_stream_ppmd_read(void *stream, void *buf, int32_t size) {
 #endif
 }
 
-int32_t mz_stream_ppmd_write(void *stream, const void *buf, int32_t size) {
+int64_t mz_stream_ppmd_write(void *stream, const void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_COMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -390,7 +390,7 @@ int64_t mz_stream_ppmd_tell(void *stream) {
     return ppmd->total_in;
 }
 
-int32_t mz_stream_ppmd_seek(void *stream, int64_t offset, int32_t origin) {
+int64_t mz_stream_ppmd_seek(void *stream, int64_t offset, int32_t origin) {
     MZ_UNUSED(stream);
     MZ_UNUSED(offset);
     MZ_UNUSED(origin);
