+++ mz_strm_zlib.h
@@ -19,10 +19,10 @@ extern "C" {
 
 int32_t mz_stream_zlib_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_zlib_is_open(void *stream);
-int32_t mz_stream_zlib_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_zlib_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_zlib_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_zlib_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_zlib_tell(void *stream);
-int32_t mz_stream_zlib_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_zlib_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_zlib_close(void *stream);
 int32_t mz_stream_zlib_error(void *stream);
