+++ mz_strm_zstd.c
@@ -85,7 +85,7 @@ int32_t mz_stream_zstd_is_open(void *stream) {
     return MZ_OK;
 }
 
-int32_t mz_stream_zstd_read(void *stream, void *buf, int32_t size) {
+int64_t mz_stream_zstd_read(void *stream, void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_DECOMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -204,7 +204,7 @@ static int32_t mz_stream_zstd_compress(void *stream, ZSTD_EndDirective flush) {
 }
 #endif
 
-int32_t mz_stream_zstd_write(void *stream, const void *buf, int32_t size) {
+int64_t mz_stream_zstd_write(void *stream, const void *buf, int64_t size) {
 #ifdef MZ_ZIP_NO_COMPRESSION
     MZ_UNUSED(stream);
     MZ_UNUSED(buf);
@@ -234,7 +234,7 @@ int64_t mz_stream_zstd_tell(void *stream) {
     return MZ_TELL_ERROR;
 }
 
-int32_t mz_stream_zstd_seek(void *stream, int64_t offset, int32_t origin) {
+int64_t mz_stream_zstd_seek(void *stream, int64_t offset, int32_t origin) {
     MZ_UNUSED(stream);
     MZ_UNUSED(offset);
     MZ_UNUSED(origin);
