+++ mz_strm_zstd.h
@@ -20,10 +20,10 @@ extern "C" {
 
 int32_t mz_stream_zstd_open(void *stream, const char *filename, int32_t mode);
 int32_t mz_stream_zstd_is_open(void *stream);
-int32_t mz_stream_zstd_read(void *stream, void *buf, int32_t size);
-int32_t mz_stream_zstd_write(void *stream, const void *buf, int32_t size);
+int64_t mz_stream_zstd_read(void *stream, void *buf, int64_t size);
+int64_t mz_stream_zstd_write(void *stream, const void *buf, int64_t size);
 int64_t mz_stream_zstd_tell(void *stream);
-int32_t mz_stream_zstd_seek(void *stream, int64_t offset, int32_t origin);
+int64_t mz_stream_zstd_seek(void *stream, int64_t offset, int32_t origin);
 int32_t mz_stream_zstd_close(void *stream);
 int32_t mz_stream_zstd_error(void *stream);
